# sync: don't parse unneeded `canonical-data.json` files